# documentation/api/swagger.yaml
openapi: 3.0.0
info:
  title: PulsePlus Game Platform API
  version: 1.0.0
  description: API documentation for the PulsePlus gaming and engagement platform
servers:
  - url: /api
    description: Main API server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    TooManyRequests:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
...
    Season:
      type: object
      properties:
        sys_id:
          type: string
          format: uuid
        name:
          type: string
        game_id:
          type: string
          format: uuid
        start_date:
          type: string
          format: date-time
        end_date:
          type: string
          format: date-time
        banner_url:
          type: string
        status:
          type: string
          enum: [UPCOMING, ACTIVE, ENDED]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    SeasonTier:
      type: object
      properties:
        sys_id:
          type: string
          format: uuid
        season_id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        level:
          type: integer
        xp_required:
          type: integer
        reward_image:
          type: string
        rewards:
          type: object
...
paths:
...
  /seasons:
    post:
      tags:
        - Season Management
      summary: Create a new season
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Season'
      responses:
        '201':
          description: Season created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Season'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /seasons/{seasonId}:
    get:
      tags:
        - Season Management
      summary: Get season details
      security:
        - BearerAuth: []
      parameters:
        - name: seasonId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Season details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Season'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'

    put:
      tags:
        - Season Management
      summary: Update season details
      security:
        - BearerAuth: []
      parameters:
        - name: seasonId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Season'
      responses:
        '200':
          description: Season updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Season'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /seasons/{seasonId}/end:
    post:
      tags:
        - Season Management
      summary: End a season and process final rankings
      security:
        - BearerAuth: []
      parameters:
        - name: seasonId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Season ended successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Season'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /seasons/tiers:
    post:
      tags:
        - Season Management
      summary: Create a new season tier
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SeasonTier'
      responses:
        '201':
          description: Season tier created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeasonTier'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'

  # Economy Management
...
security:
  - BearerAuth: [] 